+++ Game/src/GameApp/GameApp.cc
@@ -4,8 +4,17 @@
 
 using namespace Lorr;
 
+Audio *audio = new Audio;
+AudioChannel *channel = new AudioChannel;
+
 void GameApp::Init()
 {
+    AudioData data;
+    GetEngine()->GetResourceMan()->ImportAudioData("theme.lr", data);
+
+    GetEngine()->GetAudioSys()->CreateChannel("", channel);
+    GetEngine()->GetAudioSys()->LoadAudio("", audio, channel, &data);
+    audio->Play();
 }
 
 void GameApp::Tick(float fDelta)